requirements.txt
requests==2.28.2
python-dotenv==0.9.1
pytest==7.2.1
